refactor: moved repo and handler of employee to employee folder
