+++ internal/repository/employee/employee_interface.go
