+++ internal/router/employees.go
@@ -2,7 +2,7 @@ package router
 
 import (
 	"github.com/go-chi/chi/v5"
-	"github.com/varobledo_meli/W17-G10-Bootcamp.git/internal/handler"
+	handler "github.com/varobledo_meli/W17-G10-Bootcamp.git/internal/handler/employee"
 )
 
 // Monta las rutas relacionadas con la entidad Employee en el router API principal.
