+++ internal/service/inbound_order.go
@@ -5,6 +5,7 @@ import (
 	"errors"
 
 	"github.com/varobledo_meli/W17-G10-Bootcamp.git/internal/repository"
+	empRepo "github.com/varobledo_meli/W17-G10-Bootcamp.git/internal/repository/employee"
 	"github.com/varobledo_meli/W17-G10-Bootcamp.git/internal/validators"
 	"github.com/varobledo_meli/W17-G10-Bootcamp.git/pkg/api/apperrors"
 	models "github.com/varobledo_meli/W17-G10-Bootcamp.git/pkg/models/inbound_order"
@@ -13,12 +14,15 @@ import (
 // Servicio principal para Inbound Orders, incluye lógica de validación
 type InboundOrderDefault struct {
 	repo          repository.InboundOrderRepository
-	employeeRepo  repository.EmployeeRepository
+	employeeRepo  empRepo.EmployeeRepository
 	warehouseRepo repository.WarehouseRepository
 }
 
 // Constructor del servicio de inbound orders
-func NewInboundOrderService(r repository.InboundOrderRepository, er repository.EmployeeRepository, wr repository.WarehouseRepository) *InboundOrderDefault {
+func NewInboundOrderService(
+	r repository.InboundOrderRepository,
+	er empRepo.EmployeeRepository,
+	wr repository.WarehouseRepository) *InboundOrderDefault {
 	return &InboundOrderDefault{
 		repo:          r,
 		employeeRepo:  er,
